+++ src/shared/db/ign/ign.repository.ts
@@ -3,28 +3,29 @@ import { InjectModel } from '@nestjs/mongoose';
 import { validate } from 'bycontract';
 import _ from 'lodash';
 import { Model } from 'mongoose';
-import { CardStats } from './card-stats.schema';
+import { Ign } from './ign.schema';
+
 @Injectable()
-export class CardStatsRepository {
+export class IgnRepository {
   constructor(
-    @InjectModel(CardStats.name)
-    public cardModel: Model<CardStats>,
+    @InjectModel(Ign.name)
+    public ignModel: Model<Ign>,
   ) {}
 
-  async findAll(): Promise<CardStats[]> {
-    const results = await this.cardModel.find().exec();
+  async findAll(): Promise<Ign[]> {
+    const results = await this.ignModel.find().exec();
 
     return results ?? [];
   }
 
-  async save(cards: CardStats[]): Promise<void> {
+  async save(cards: Ign[]): Promise<void> {
     validate([cards], ['Array.<object>']);
 
     if (cards.length === 0) {
       return;
     }
 
-    await this.cardModel.bulkWrite(
+    await this.ignModel.bulkWrite(
       cards.map((model) => {
         validate(model, 'object');
         return {
@@ -33,21 +34,16 @@ export class CardStatsRepository {
               id: model.id,
             },
             update: {
-              $set: _.pick(model, [
-                'id',
-                'blockNumber',
-                'hash',
-                'level',
-                'dailyBattleStats',
-                'templateId',
-                'gold',
-                'editionNumber',
-              ]),
+              $set: _.pick(model, ['id']),
             },
             upsert: true,
           },
         };
       }),
     );
+  }
+
+  async delete(id: string) {
+    await this.ignModel.deleteOne({ id });
   }
 }