+++ src/shared/db/ign/index.ts
@@ -1,0 +1,2 @@
+export * from './ign.repository';
+export * from './ign.schema';