+++ src/shared/db/index.ts
@@ -1,3 +1,4 @@
 export * from './battle';
 export * from './card';
 export * from './db.module';
+export * from './ign';